analyzer-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
google-generativeai==0.3.2
beautifulsoup4==4.12.2
aiohttp==3.9.1